requirements.txt
Pillow==11.1.0
requests==2.32.3
retry==0.9.2
